cli: Move EpochTime to cliutil
